refactor: extract rolling dependency context
